chore: Quizzing Changes Before Migration

+++ QuizzArena.Backend/Modules/QuizzArena.Quizzing/Domain/Entities/MatchAttemptQuestion.cs
@@ -4,7 +4,7 @@ namespace QuizzArena.Quizzing.Domain.Entities;
 public class MatchAttemptQuestion
 {
     public Guid Id { get; set; }
-    public int? ValueScore { get; set; }
+    public decimal? ValueScore { get; set; }
     public Guid MatchAttemptId { get; set; }
     public Guid QuestionId { get; set; }
