+++ QuizzArena.Backend/Modules/QuizzArena.Quizzing/Domain/Entities/QuizQuestion.cs
@@ -4,7 +4,7 @@ public class QuizQuestion
 {
     public Guid Id { get; set; }
     public int Position { get; set; }
-    public int ValueScore { get; set; }
+    public decimal ValueScore { get; set; }
     public bool Deleted { get; set; }
     public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow;
     public DateTimeOffset UpdatedAt { get; set; } = DateTimeOffset.UtcNow;
